requirements.txt
prettytable>=3.7.0
python-nmap>=0.7.1
colorama>=0.4.6
dataclasses>=0.6; python_version < '3.7'
pyinstaller>=6.5.0
python-json-logger>=2.0.7  # Para formatação JSON dos logs
PyYAML>=6.0.1  # Para o sistema de configuração

# Development dependencies
pytest>=7.4.0
pytest-mock>=3.11.1
pytest-cov>=4.1.0